backend/ArchieMCPFunctionApp/requirements.txt
azure-functions
openai>=2.0.0 # Or specify the version you prefer
# Add any other dependencies your function might need below